added checkpoint config, fixed unit tests, updated checkpoint related implementation, renamed namespace for convinience
