+++ src/Lachain.Console/Application.cs
@@ -14,6 +14,7 @@ using Lachain.Core.Network;
 using Lachain.Core.RPC;
 using Lachain.Core.ValidatorStatus;
 using Lachain.Core.Vault;
+using Lachain.Core.Network.FastSync;
 using Lachain.Crypto;
 using Lachain.Logger;
 using Lachain.Networking;
