+++ src/Lachain.Core/Network/BlockSynchronizer.cs
@@ -11,7 +11,7 @@ using Lachain.Core.Blockchain.Hardfork;
 using Lachain.Core.Blockchain.Interface;
 using Lachain.Core.Blockchain.Pool;
 using Lachain.Core.Consensus;
-using Lachain.Core.Network.FastSynchronizerBatch;
+using Lachain.Core.Network.FastSync;
 using Lachain.Logger;
 using Lachain.Networking;
 using Lachain.Proto;
@@ -398,7 +398,7 @@ namespace Lachain.Core.Network
         // Here we decide if we need to do fast sync
         private bool IsFastSyncNeeded(ulong myHeight, ulong maxHeight)
         {
-            if (myHeight + FastSynchronizerBatch.FastSynchronizerBatch.FastSyncBlockDiff > maxHeight) return false;
+            if (myHeight + FastSynchronizerBatch.FastSyncBlockDiff > maxHeight) return false;
             CheckIfCheckpointExist();
             lock (_peerHasCheckpoint)
             {
@@ -407,8 +407,8 @@ namespace Lachain.Core.Network
             GetCheckpoint();
             lock (_peerHasCheckpoint)
             {
-                if (_checkpointBlockHash is null) return false;
-                return myHeight + FastSynchronizerBatch.FastSynchronizerBatch.FastSyncBlockDiff <= _checkpointBlockHeight;
+                if (_checkpointBlockHeight is null) return false;
+                return myHeight + FastSynchronizerBatch.FastSyncBlockDiff <= _checkpointBlockHeight;
             }
         }
