+++ src/Lachain.Core/Network/FastSync/BlockRequestManager.cs
@@ -14,7 +14,7 @@ using Lachain.Proto;
 using Lachain.Utility.Utils;
 
 
-namespace Lachain.Core.Network.FastSynchronizerBatch
+namespace Lachain.Core.Network.FastSync
 {
     public class BlockRequestManager : IBlockRequestManager
     {
