+++ src/Lachain.Core/Network/FastSync/Downloader.cs
@@ -15,7 +15,7 @@ using Lachain.Proto;
 using Lachain.Utility.Utils;
 
 
-namespace Lachain.Core.Network.FastSynchronizerBatch
+namespace Lachain.Core.Network.FastSync
 {
     public class Downloader : IDownloader
     {
