+++ src/Lachain.Core/Network/FastSync/FastSyncRepository.cs
@@ -20,7 +20,7 @@ using Lachain.Utility.Serialization;
 using Lachain.Utility.Utils;
 
 
-namespace Lachain.Core.Network.FastSynchronizerBatch
+namespace Lachain.Core.Network.FastSync
 {
     public class FastSyncRepository : IFastSyncRepository
     {
