+++ src/Lachain.Core/Network/FastSync/IFastSynchronizerBatch.cs
@@ -3,7 +3,7 @@ using Lachain.Core.Blockchain.Checkpoints;
 using Lachain.Proto;
 
 
-namespace Lachain.Core.Network.FastSynchronizerBatch
+namespace Lachain.Core.Network.FastSync
 {
     public interface IFastSynchronizerBatch
     {
