+++ src/Lachain.Core/Network/FastSync/IRequestManager.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using Lachain.Proto;
 
 
-namespace Lachain.Core.Network.FastSynchronizerBatch
+namespace Lachain.Core.Network.FastSync
 {
     public interface IRequestManager
     {
