+++ src/Lachain.Core/Network/FastSync/Peer.cs
@@ -2,7 +2,7 @@ using System;
 using Lachain.Proto;
 
 
-namespace Lachain.Core.Network.FastSynchronizerBatch
+namespace Lachain.Core.Network.FastSync
 {
     public class Peer : IEquatable<Peer>
     {
