+++ src/Lachain.Core/Network/FastSync/PeerManager.cs
@@ -9,7 +9,7 @@ using System.Runtime.CompilerServices;
 using Lachain.Proto;
 
 
-namespace Lachain.Core.Network.FastSynchronizerBatch
+namespace Lachain.Core.Network.FastSync
 {
     public class PeerManager
     {
