+++ src/Lachain.Core/Network/FastSync/RequestManager.cs
@@ -11,7 +11,7 @@ using Lachain.Storage.Trie;
 using Lachain.Utility.Utils;
 
 
-namespace Lachain.Core.Network.FastSynchronizerBatch
+namespace Lachain.Core.Network.FastSync
 {
     public class RequestManager : IRequestManager
     {
