+++ src/Lachain.Core/Network/FastSync/RequestState.cs
@@ -5,7 +5,7 @@ using Lachain.Proto;
 using Lachain.Utility.Serialization;
 
 
-namespace Lachain.Core.Network.FastSynchronizerBatch
+namespace Lachain.Core.Network.FastSync
 {
     public class RequestState
     {
